Refactor: upload에서 objectstorage로 폴더 이름 변경
