+++ src/main/java/com/sppart/admin/exhibition/service/ExhibitionServiceImpl.java
@@ -13,9 +13,9 @@ import com.sppart.admin.exhibition.dto.ResponseExhibitionByCondition;
 import com.sppart.admin.exhibition.dto.ResponseGetExhibitionsByCondition;
 import com.sppart.admin.exhibition.dto.request.RequestCreateExhibition;
 import com.sppart.admin.exhibition.exception.ExhibitionErrorCode;
+import com.sppart.admin.objectstorage.service.ObjectStorageService;
 import com.sppart.admin.product.dto.ProductOnlyArtistNameDto;
 import com.sppart.admin.productexhibition.mapper.ProductExhibitionMapper;
-import com.sppart.admin.upload.service.ObjectStorageService;
 import java.util.List;
 import java.util.Set;
 import java.util.function.Predicate;
