+++ src/test/java/com/sppart/admin/exhibition/service/ExhibitionServiceImplTest.java
@@ -13,8 +13,8 @@ import com.sppart.admin.exhibition.domain.mapper.ExhibitionMapper;
 import com.sppart.admin.exhibition.dto.ExhibitionSearchCondition;
 import com.sppart.admin.exhibition.dto.RequestUpdateExhibitionDisplay;
 import com.sppart.admin.exhibition.dto.request.RequestCreateExhibition;
+import com.sppart.admin.objectstorage.service.ObjectStorageService;
 import com.sppart.admin.productexhibition.mapper.ProductExhibitionMapper;
-import com.sppart.admin.upload.service.ObjectStorageService;
 import java.nio.charset.StandardCharsets;
 import java.time.LocalDate;
 import java.util.HashSet;
